# api-gateway/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  version: 1.0.0
  description: Public API served through the API Gateway. Internal service endpoints are intentionally excluded.
servers:
  - url: /
security: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Manual login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access token or 2FA pending token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/google:
    post:
      tags: [Auth]
      summary: Google ID-token login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token:
                  type: string
                  maxLength: 4096
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh cookie and issue a new access token
      responses:
        "200":
          description: New access token and refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Error"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current refresh-cookie session
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "429":
          $ref: "#/components/responses/RateLimited"
        "401":
          $ref: "#/components/responses/Error"
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: Logout all sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    OK:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
...
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limited
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Success:
      type: object
      required: [success, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorResponse:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
            retry_after_seconds:
              type: integer
        meta:
          $ref: "#/components/schemas/Meta"
    Meta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          maxLength: 255
        password:
          type: string
          maxLength: 128
          format: password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        access_token_expires_at:
          type: string
          format: date-time
        two_factor_required:
          type: boolean
        two_factor_session_token:
          type: string
...